(fix): Move the creation subject button to another place - request of ODE
