+++ admin/src/main/ts/src/app/management/management-root/management-root.component.ts
@@ -16,6 +16,7 @@ export class ManagementRootComponent extends OdeComponent implements OnInit, OnD
         { label: 'management.message.flash', view: 'message-flash/list', active: 'message-flash'},
         { label: 'management.block.profile.tab', view: 'block-profiles', active: 'block-profiles'},
         { label: 'management.edt.tab', view: 'import-edt', active: 'import-edt'},
+        { label: 'create.subjects', view: 'subjects/create', active: 'subjects'},
     ];
 
     private structure: StructureModel;
