+++ admin/src/main/ts/src/app/management/subjects/subjects.component.html
@@ -3,13 +3,13 @@
         <s5l>{{ 'subjects' | translate}}</s5l>
     </h1>
     <button [routerLink]="['create']"
-            [class.hidden]="router.isActive('/admin/' + subjectsStore.structure?.id + '/subjects/create', true)">
+            [class.hidden]="router.isActive('/admin/' + subjectsStore.structure?.id + '/management/subjects/create', true)">
         <s5l>create.subjects</s5l>
         <i class="fa fa-plus-circle is-size-5"></i>
     </button>
 </div>
 <ode-side-layout (closeCompanion)="closeCompanion()"
-[showCompanion]="!router.isActive('/admin/' + subjectsStore.structure?.id + '/subjects', true)">
+[showCompanion]="!router.isActive('/admin/' + subjectsStore.structure?.id + '/management/subjects', true)">
 <div side-card>
     <ode-list
             [model]="subjectsStore.structure?.subjects.data"
