+++ admin/src/main/ts/src/app/structure/cards/quick-actions-card/quick-actions-card.component.html
@@ -21,8 +21,4 @@
     <s5l>create.flash.message</s5l>
     <i class="fa fa-bullhorn"></i>
   </button>
-  <button routerLink="subjects/create">
-    <s5l>create.subjects</s5l>
-    <i class="fa fa-file"></i>
-  </button>
 </div>
