+++ admin/src/main/ts/src/app/structure/structure.routing.ts
@@ -13,8 +13,7 @@ export let routes: Routes = [
 			{ path: 'imports-exports', loadChildren: () => import('../imports-exports/imports-exports.module').then(m => m.ImportsExportsModule)},
 			{ path: 'services', loadChildren: () => import('../services/services.module').then(m => m.ServicesModule)},
 			{ path: 'management', loadChildren: () => import('../management/management.module').then(m => m.ManagementModule)},
-			{path: 'alertes', loadChildren: () => import('../alertes/alertes.module').then(m => m.AlertesModule)},
-			{path: 'subjects', loadChildren: () => import('../subjects/subjects.module').then(m => m.SubjectsModule)}
-		]
+			{ path: 'alertes', loadChildren: () => import('../alertes/alertes.module').then(m => m.AlertesModule)}
+			]
 	}
 ];
